# api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoCertX Control Plane and Agent API",
    "version": "1.0.1-t03",
    "description": "Initial frozen API contract for GA phase one. Covers control plane REST APIs, agent REST protocol, and shared error semantics."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "operationId": "login",
        "summary": "Login with username and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthLoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthLoginResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "operationId": "refreshToken",
        "summary": "Refresh access token with refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthLoginResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["auth"],
        "operationId": "logout",
        "summary": "Logout current session",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmptyEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["auth"],
        "operationId": "getAuthContext",
        "summary": "Get current user context",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user context",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthMeResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/auth/me/preferences": {
      "patch": {
        "tags": ["auth"],
        "operationId": "updateUserPreferences",
        "summary": "Update current user preferences",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LocalePreferencesRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user context",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthMeResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": { "type": "string" },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "field": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
...
      "EmptyEnvelope": {
        "type": "object",
        "required": ["request_id", "status"],
        "properties": {
          "request_id": { "type": "string" },
          "status": { "type": "string" }
        }
      },
      "AuthLoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "AuthLoginResponse": {
        "type": "object",
        "required": ["request_id", "access_token", "refresh_token", "expires_in", "context"],
        "properties": {
          "request_id": { "type": "string" },
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "expires_in": { "type": "integer", "minimum": 1 },
          "context": { "$ref": "#/components/schemas/AuthContext" }
        }
      },
      "AuthRefreshRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": { "type": "string" }
        }
      },
      "AuthMeResponse": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": { "type": "string" },
          "data": { "$ref": "#/components/schemas/AuthContext" }
        }
      },
      "AuthContext": {
        "type": "object",
        "required": ["locale", "available_locales", "user", "tenant", "project", "environment", "roles"],
        "properties": {
          "locale": { "type": "string", "enum": ["zh-CN", "en-US"] },
          "available_locales": {
            "type": "array",
            "items": { "type": "string", "enum": ["zh-CN", "en-US"] }
          },
          "user": { "$ref": "#/components/schemas/NamedResource" },
          "tenant": { "$ref": "#/components/schemas/NamedResource" },
          "project": { "$ref": "#/components/schemas/NamedResource" },
          "environment": { "$ref": "#/components/schemas/NamedResource" },
          "roles": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "NamedResource": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "code": { "type": "string" },
          "display_name": { "type": "string" }
        }
      },
      "LocalePreferencesRequest": {
        "type": "object",
        "required": ["locale"],
        "properties": {
          "locale": { "type": "string", "enum": ["zh-CN", "en-US"] }
        }
      },
...
    }
  }
}